+++ src/sankeyDiagram.ts
@@ -27,34 +27,35 @@ import "../style/visual.less";
 import powerbi from "powerbi-visuals-api";
 
 // lodash
-import * as _ from "lodash-es";
+import lodashMaxBy from "lodash.maxby";
+import lodashMinBy from "lodash.minby";
+import lodashUniq from "lodash.uniq";
+import lodashCloneDeep from "lodash.clonedeep";
+
 // d3
-import * as d3 from "d3";
+import { select as d3Select, Selection as d3Selection } from "d3-selection";
+import { drag as d3Drag } from "d3-drag";
+import { max as d3Max, min as d3Min } from "d3-array";
+import { scaleLog as d3ScaleLog, scaleLinear as d3ScaleLinear } from "d3-scale";
+import { rgb as d3Rgb } from "d3-color";
+import { interpolateNumber as d3InterpolateNumber } from "d3-interpolate";
 
-const getEvent = () => <MouseEvent>require("d3-selection").event;
-
-type Selection<T> = d3.Selection<any, T, any, any>;
-type UpdateSelection<T> = d3.Selection<any, T, any, any>;
+type Selection<T> = d3Selection<any, T, any, any>;
+type UpdateSelection<T> = d3Selection<any, T, any, any>;
 
 // powerbi
 import DataView = powerbi.DataView;
 import IViewport = powerbi.IViewport;
 import DataViewObjects = powerbi.DataViewObjects;
-import DataViewValueColumn = powerbi.DataViewValueColumn;
 import VisualObjectInstance = powerbi.VisualObjectInstance;
-import DataViewCategoryColumn = powerbi.DataViewCategoryColumn;
 import VisualObjectInstanceEnumeration = powerbi.VisualObjectInstanceEnumeration;
 import DataViewObjectPropertyIdentifier = powerbi.DataViewObjectPropertyIdentifier;
 import EnumerateVisualObjectInstancesOptions = powerbi.EnumerateVisualObjectInstancesOptions;
 import IVisualHost = powerbi.extensibility.visual.IVisualHost;
 import ILocalizationManager = powerbi.extensibility.ILocalizationManager;
-import DataViewMetadataColumn = powerbi.DataViewMetadataColumn;
-import DataViewValueColumns = powerbi.DataViewValueColumns;
 import VisualObjectInstanceEnumerationObject = powerbi.VisualObjectInstanceEnumerationObject;
 import DataViewMatrixNode = powerbi.DataViewMatrixNode;
-import DataViewMatrix = powerbi.DataViewMatrix;
 // powerbi.visuals
-import ISelectionIdBuilder = powerbi.visuals.ISelectionIdBuilder;
 import ISelectionId = powerbi.visuals.ISelectionId;
 
 // powerbi.extensibility
@@ -82,7 +83,7 @@ import TextProperties = interfaces.TextProperties;
 import IValueFormatter = valueFormatter.IValueFormatter;
 
 // powerbi.extensibility.utils.interactivity
-import { interactivitySelectionService, interactivityUtils, interactivityBaseService } from "powerbi-visuals-utils-interactivityutils";
+import { interactivitySelectionService, interactivityBaseService } from "powerbi-visuals-utils-interactivityutils";
 import appendClearCatcher = interactivityBaseService.appendClearCatcher;
 import SelectableDataPoint = interactivitySelectionService.SelectableDataPoint;
 import IInteractiveBehavior = interactivityBaseService.IInteractiveBehavior;
@@ -91,7 +92,6 @@ import createInteractivitySelectionService = interactivitySelectionService.creat
 
 // powerbi.extensibility.utils.tooltip
 import {
-    TooltipEventArgs,
     ITooltipServiceWrapper,
     TooltipEnabledDataPoint,
     createTooltipServiceWrapper
@@ -111,7 +111,6 @@ import {
 import {
     SankeyDiagramColumn,
     SankeyDiagramCycleDictionary,
-    SankeyDiagramDataPoint,
     SankeyDiagramDataView,
     SankeyDiagramLabel,
     SankeyDiagramLink,
@@ -132,7 +131,6 @@ import {
     SankeyDiagramBehaviorOptions,
     SankeyDiagramBehavior
 } from "./behavior";
-import { data } from "jquery";
 
 export class SankeyDiagram implements IVisual {
     private static ClassName: string = "sankeyDiagram";
@@ -284,9 +282,6 @@ export class SankeyDiagram implements IVisual {
     }
 
     constructor(options: VisualConstructorOptions) {
-        if (window.location !== window.parent.location) {
-            require("core-js/stable");
-        }
         this.init(options);
     }
 
@@ -294,7 +289,7 @@ export class SankeyDiagram implements IVisual {
         this.visualHost = options.host;
         this.localizationManager = this.visualHost.createLocalizationManager();
 
-        this.root = d3.select(options.element)
+        this.root = d3Select(options.element)
             .append("svg")
             .classed(SankeyDiagram.ClassName, true);
 
@@ -324,18 +319,14 @@ export class SankeyDiagram implements IVisual {
 
     public update(visualUpdateOptions: VisualUpdateOptions): void {
         this.visualHost.eventService.renderingStarted(visualUpdateOptions);
-        
-        let sankeyDiagramDataView: SankeyDiagramDataView,
-            viewport: IViewport = visualUpdateOptions
-                && visualUpdateOptions.viewport
-                || SankeyDiagram.DefaultViewport,
-            dataView: DataView = visualUpdateOptions
+
+        const dataView: DataView = visualUpdateOptions
                 && visualUpdateOptions.dataViews
                 && visualUpdateOptions.dataViews[0];
-        
+
         this.updateViewport(visualUpdateOptions.viewport);
 
-        sankeyDiagramDataView = this.converter(dataView);
+        const sankeyDiagramDataView: SankeyDiagramDataView = this.converter(dataView);
 
         this.computePositions(sankeyDiagramDataView);
 
@@ -345,15 +336,12 @@ export class SankeyDiagram implements IVisual {
 
         this.render(sankeyDiagramDataView);
         this.visualHost.eventService.renderingFinished(visualUpdateOptions);
-       
+
     }
 
     private updateViewport(viewport: IViewport): void {
-        let height: number,
-            width: number;
-
-        height = SankeyDiagram.getPositiveNumber(viewport.height);
-        width = SankeyDiagram.getPositiveNumber(viewport.width);
+        const height: number = SankeyDiagram.getPositiveNumber(viewport.height);
+        const width: number = SankeyDiagram.getPositiveNumber(viewport.width);
 
         this.viewport = {
             height: SankeyDiagram.getPositiveNumber(height - this.margin.top - this.margin.bottom),
@@ -379,20 +367,20 @@ export class SankeyDiagram implements IVisual {
     }
 
     private createNewNode(node: DataViewMatrixNode, settings: SankeyDiagramSettings): SankeyDiagramNode {
-        let nodeFillColor = this.getColor(
+        const nodeFillColor = this.getColor(
             SankeyDiagram.NodesPropertyIdentifier,
             this.colorPalette.getColor(<string>node.value).value,
             <any>node.objects);
-        let nodeStrokeColor = this.colorHelper.getHighContrastColor("foreground", nodeFillColor);
+        const nodeStrokeColor = this.colorHelper.getHighContrastColor("foreground", nodeFillColor);
 
-        let name = <any>node.value;
+        const name = <any>node.value;
 
-        let textProperties: TextProperties = {
+        const textProperties: TextProperties = {
             text: name,
             fontFamily: this.textProperties.fontFamily,
             fontSize: this.textProperties.fontSize
         };
-        let label: SankeyDiagramLabel = {
+        const label: SankeyDiagramLabel = {
             internalName: name,
             name: name,
             formattedName: name,//valueFormatterForCategories.format((<string>labelsDictionary[item].toString()).replace(SankeyDiagram.DuplicatedNamePostfix, "")),
@@ -420,7 +408,7 @@ export class SankeyDiagram implements IVisual {
         }
     }
 
-    // tslint:disable-next-line: max-func-body-length
+    /*eslint max-lines-per-function: ["error", 200]*/
     public converter(dataView: DataView): SankeyDiagramDataView {
         const settings: SankeyDiagramSettings = this.parseSettings(dataView);
 
@@ -447,26 +435,13 @@ export class SankeyDiagram implements IVisual {
             }
         }
 
-        let nodes: SankeyDiagramNode[] = [],
-            links: SankeyDiagramLink[] = [],
-            categories: any[] = [],
-            sourceCategories: any[] = [],
-            destinationCategories: any[] = [],
-            sourceCategoriesLabels: any[] = [],
-            destinationCategoriesLabels: any[] = [],
-            objects: any[] = [],
-            weights: any[] = [],
-            selectionIdBuilder: SelectionIdBuilder = new SelectionIdBuilder(
-                this.visualHost,
-                dataView.matrix),
-            valueSources = dataView.matrix.valueSources;
+        const nodes: SankeyDiagramNode[] = [];
+        let links: SankeyDiagramLink[] = [];
 
-
-
+        const valueSources = dataView.matrix.valueSources;
         const sourceLabelIndex: number = valueSources.indexOf(valueSources.filter((column: powerbi.DataViewMetadataColumn) => {
             return column.roles.SourceLabels;
         }).pop());
-
 
         const weightIndex: number = valueSources.indexOf(valueSources.filter((source: powerbi.DataViewMetadataColumn) => {
             return source.roles.Weight;
@@ -475,14 +450,11 @@ export class SankeyDiagram implements IVisual {
         const sourceFieldName = dataView.matrix.rows.levels[0].sources[0].displayName;
         const destinationFieldName = dataView.matrix.rows.levels[1].sources[0].displayName;
         const valueFieldName = dataView.matrix.valueSources[weightIndex] ? dataView.matrix.valueSources[weightIndex].displayName : null;
-
         const formatOfWeigth = valueFormatter.getFormatStringByColumn(valueSources[weightIndex]);
-
-
-        let weightValues: number[] = [1];
+        const weightValues: number[] = [1];
 
         dataView.matrix.rows.root.children.forEach(parent => {
-            let newSourceNode = this.createNewNode(parent, settings)
+            const newSourceNode = this.createNewNode(parent, settings)
             newSourceNode.identity = this.visualHost.createSelectionIdBuilder()
                 .withMatrixNode(parent, dataView.matrix.rows.levels)
                 .createSelectionId();
@@ -490,10 +462,8 @@ export class SankeyDiagram implements IVisual {
 
         });
 
-
         dataView.matrix.rows.root.children.forEach(parent => {
-            let foundSource: SankeyDiagramNode = nodes.find(found => found.label.name === parent.value)
-
+            const foundSource: SankeyDiagramNode = nodes.find(found => found.label.name === parent.value)
             parent.children.forEach(child => {
                 let linkLabel = undefined;
                 let weigth: any = SankeyDiagram.DefaultWeightValue;
@@ -507,31 +477,31 @@ export class SankeyDiagram implements IVisual {
                         .createSelectionId();
                     nodes.push(foundDestination);
                 }
-
                 if (sourceLabelIndex != -1) {
                     linkLabel = (child.values[sourceLabelIndex] && child.values[sourceLabelIndex].value) ?
                         child.values[sourceLabelIndex].value || SankeyDiagram.DefaultWeightValue : SankeyDiagram.MinWeightValue;
                 }
-
-
                 // If weights are present, populate the weights array
                 if (weightIndex != -1) {
                     weigth = (child.values[weightIndex] && child.values[weightIndex].value) ?
                         child.values[weightIndex].value || SankeyDiagram.DefaultWeightValue : SankeyDiagram.MinWeightValue;
-
                     weightValues.push(weigth);
-
                 }
-
-                let linkFillColor = this.getColor(
+                const linkFillColor = this.getColor(
                     SankeyDiagram.LinksPropertyIdentifier,
                     SankeyDiagram.DefaultColourOfLink,
                     child.objects);
-                let linkStrokeColor = this.colorHelper.isHighContrast ? this.colorHelper.getHighContrastColor("foreground", linkFillColor) : linkFillColor;
+                const linkStrokeColor = this.colorHelper.isHighContrast ? this.colorHelper.getHighContrastColor("foreground", linkFillColor) : linkFillColor;
 
+                const valuesFormatterForLinkTooltipInfo = valueFormatter.create({
+                    format: formatOfWeigth,
+                    value: Math.max(
+                        settings.labels.unit !== 0 ? settings.labels.unit : d3Max(weightValues) || SankeyDiagram.MinWeightValue,
+                        SankeyDiagram.MinWeightValue),
+                });
 
-                let tooltipInfo = SankeyDiagram.getTooltipDataForLink(
-                    valuesFormatterForWeigth,
+                const tooltipInfo = SankeyDiagram.getTooltipDataForLink(
+                    valuesFormatterForLinkTooltipInfo,
                     foundSource.label.formattedName,
                     foundDestination.label.formattedName,
                     weigth,
@@ -540,7 +510,7 @@ export class SankeyDiagram implements IVisual {
                     valueFieldName
                 );
 
-                let link: SankeyDiagramLink = {
+                const link: SankeyDiagramLink = {
                     label: linkLabel && linkLabel.toString(),
                     source: foundSource,
                     destination: foundDestination,
@@ -559,15 +529,12 @@ export class SankeyDiagram implements IVisual {
                     direction: SankeyLinkDirrections.Forward
                 }
 
-                let selectableDataPoint: SelectableDataPoint = SankeyDiagram.createSelectableDataPoint(<ISelectionId>link.identity);
+                const selectableDataPoint: SelectableDataPoint = SankeyDiagram.createSelectableDataPoint(<ISelectionId>link.identity);
                 foundSource.selectableDataPoints.push(selectableDataPoint);
                 foundDestination.selectableDataPoints.push(selectableDataPoint);
-
                 links.push(link);
-
                 foundSource.links.push(link);
                 foundDestination.links.push(link);
-
                 SankeyDiagram.updateValueOfNode(foundSource);
                 SankeyDiagram.updateValueOfNode(foundDestination);
             });
@@ -576,12 +543,11 @@ export class SankeyDiagram implements IVisual {
         const valuesFormatterForWeigth = valueFormatter.create({
             format: formatOfWeigth,
             value: Math.max(
-                settings.labels.unit !== 0 ? settings.labels.unit : d3.max(weightValues) || SankeyDiagram.MinWeightValue,
+                settings.labels.unit !== 0 ? settings.labels.unit : d3Max(weightValues) || SankeyDiagram.MinWeightValue,
                 SankeyDiagram.MinWeightValue),
         });
 
-
-        let cycles: SankeyDiagramCycleDictionary = this.checkCycles(nodes);
+        const cycles: SankeyDiagramCycleDictionary = this.checkCycles(nodes);
 
         if (settings.cyclesLinks.drawCycles === CyclesDrawType.Duplicate) {
             links = this.processCyclesForwardLinks(cycles, nodes, links, settings);
@@ -593,7 +559,7 @@ export class SankeyDiagram implements IVisual {
                 node.label.formattedName,
                 node.inputWeight
                     ? node.inputWeight
-                    : node.outputWeight,
+                    : this.calculeOutputWeightForTooltip(node, node.links),
                 this.localizationManager,
                 node.inputWeight > 0 && node.outputWeight > 0 ? `${sourceFieldName}-${destinationFieldName}` : node.outputWeight > 0
                     ? sourceFieldName
@@ -602,8 +568,7 @@ export class SankeyDiagram implements IVisual {
             );
         });
 
-
-        let sankeyDiagramDataView = {
+        const sankeyDiagramDataView = {
             nodes,
             links,
             settings,
@@ -624,10 +589,27 @@ export class SankeyDiagram implements IVisual {
         return sankeyDiagramDataView;
     }
 
+    // Proper calculation of output weight in case of self links
+    private calculeOutputWeightForTooltip(node: SankeyDiagramNode, links: SankeyDiagramLink[]): number {
+        let outputWeight: number = 0;
+
+        links.forEach((link: SankeyDiagramLink) => {
+            outputWeight +=
+                link.source === node &&
+                    link.destination !== link.source
+                    ?
+                    link.weigth
+                    :
+                    SankeyDiagram.DefaultWeightValue;
+        });
+
+        return outputWeight;
+    }
+
 
     private static swapNodes(link: SankeyDiagramLink) {
         link.direction = SankeyLinkDirrections.Backward;
-        let source = link.source;
+        const source = link.source;
         link.source = link.destination;
         link.destination = source;
         SankeyDiagram.updateValueOfNode(link.destination);
@@ -635,11 +617,11 @@ export class SankeyDiagram implements IVisual {
     }
 
     private processCyclesForwardLinks(cycles: SankeyDiagramCycleDictionary, nodes: SankeyDiagramNode[], links: SankeyDiagramLink[], settings: SankeyDiagramSettings): SankeyDiagramLink[] {
-        let createdNodes: SankeyDiagramNode[] = [];
-        for (let nodeName of Object.keys(cycles)) {
-            let firstCyclesNode: SankeyDiagramNode = cycles[nodeName][cycles[nodeName].length - 1];
+        const createdNodes: SankeyDiagramNode[] = [];
+        for (const nodeName of Object.keys(cycles)) {
+            const firstCyclesNode: SankeyDiagramNode = cycles[nodeName][cycles[nodeName].length - 1];
             // create a clone of the node and save a link to each other. In selection behavior, selection of clone lead to select original and visa versa
-            let nodeCopy: SankeyDiagramNode = _.cloneDeep(firstCyclesNode);
+            const nodeCopy: SankeyDiagramNode = lodashCloneDeep(firstCyclesNode);
             nodeCopy.label.name += SankeyDiagram.DuplicatedNamePostfix;
             firstCyclesNode.cloneLink = nodeCopy;
             nodeCopy.cloneLink = firstCyclesNode;
@@ -677,12 +659,11 @@ export class SankeyDiagram implements IVisual {
 
     // in this method we breaking simple cycles
     private processCyclesForBackwardLinks(cycles: SankeyDiagramCycleDictionary, nodes: SankeyDiagramNode[], links: SankeyDiagramLink[], settings: SankeyDiagramSettings): SankeyDiagramLink[] {
-        let createdNodes: SankeyDiagramNode[] = [];
-        for (let nodeName of Object.keys(cycles)) {
-            let firstCyclesNode: SankeyDiagramNode = cycles[nodeName][cycles[nodeName].length - 1];
+        for (const nodeName of Object.keys(cycles)) {
+            const firstCyclesNode: SankeyDiagramNode = cycles[nodeName][cycles[nodeName].length - 1];
 
             // make output links as backward links for node
-            let outputLinks = firstCyclesNode.links.filter((link: SankeyDiagramLink) => {
+            const outputLinks = firstCyclesNode.links.filter((link: SankeyDiagramLink) => {
                 if (link.source === firstCyclesNode || link.source === link.destination) {
                     return true;
                 }
@@ -702,10 +683,10 @@ export class SankeyDiagram implements IVisual {
     }
 
     private checkNodePositionSettings(nodes: SankeyDiagramNode[], settings: SankeyDiagramSettings) {
-        let nodePositions: SankeyDiagramNodePositionSetting[] = settings._nodePositions;
+        const nodePositions: SankeyDiagramNodePositionSetting[] = settings._nodePositions;
 
         nodePositions.forEach((position: SankeyDiagramNodePositionSetting) => {
-            let check: boolean = nodes.some((node: SankeyDiagramNode) => {
+            const check: boolean = nodes.some((node: SankeyDiagramNode) => {
                 if (node.label.name === position.name) {
                     return true;
                 }
@@ -723,7 +704,7 @@ export class SankeyDiagram implements IVisual {
 
     private restoreNodePositions(nodes: SankeyDiagramNode[], settings: SankeyDiagramSettings) {
         nodes.forEach((node: SankeyDiagramNode) => {
-            let nodeSettings: SankeyDiagramNodePositionSetting = this.getNodeSettings(node.label.name, settings);
+            const nodeSettings: SankeyDiagramNodePositionSetting = this.getNodeSettings(node.label.name, settings);
             node.settings = nodeSettings;
         });
     }
@@ -731,7 +712,7 @@ export class SankeyDiagram implements IVisual {
     // remove Duplicated links
     private static fixLinksCount(node: SankeyDiagramNode) {
         // tslint:disable-next-line: underscore-consistent-invocation
-        node.links = _.uniq(node.links);
+        node.links = lodashUniq(node.links);
     }
 
     // tslint:disable-next-line: function-name
@@ -745,7 +726,7 @@ export class SankeyDiagram implements IVisual {
             }
 
             // get node by output link
-            let nextNode: SankeyDiagramNode = link.destination;
+            const nextNode: SankeyDiagramNode = link.destination;
             // move to next not visited node
             if (nodesStatuses[nextNode.label.name].status === SankeyDiagramNodeStatus.NotVisited) {
                 SankeyDiagram.dfs(nodes, nextNode, nodesStatuses, simpleCycles);
@@ -780,7 +761,7 @@ export class SankeyDiagram implements IVisual {
 
     // Depth-First Search
     private checkCycles(nodes: SankeyDiagramNode[]): SankeyDiagramCycleDictionary {
-        let nodesStatuses: SankeyDiagramNodeStatus[] = [];
+        const nodesStatuses: SankeyDiagramNodeStatus[] = [];
 
         // init nodes statuses array
         // all nodes are not visited state
@@ -793,7 +774,7 @@ export class SankeyDiagram implements IVisual {
             }
         });
 
-        let simpleCycles: SankeyDiagramCycleDictionary = {};
+        const simpleCycles: SankeyDiagramCycleDictionary = {};
 
         nodes.forEach((node: SankeyDiagramNode) => {
             if (nodesStatuses[node.label.name].status === SankeyDiagramNodeStatus.NotVisited &&
@@ -806,7 +787,7 @@ export class SankeyDiagram implements IVisual {
     }
 
 
-    
+
     private static createSelectableDataPoint(
         selectionId: ISelectionId,
         isSelected: boolean = false): SelectableDataPoint {
@@ -863,7 +844,7 @@ export class SankeyDiagram implements IVisual {
             formattedLinkWeight = linkWeight.toString();
         }
 
-        let tooltips: VisualTooltipDataItem[] = [
+        const tooltips: VisualTooltipDataItem[] = [
             {
                 displayName: sourceNodeDisplayName || SankeyDiagram.RoleNames.rows,
                 value: sourceNodeName
@@ -918,7 +899,7 @@ export class SankeyDiagram implements IVisual {
         nodeWeight: number,
         localizationManager: ILocalizationManager,
         nodeDisplayName?: string,
-        valueDisplayName?: string
+        valueDisplayName?: string,
     ): VisualTooltipDataItem[] {
 
         let formattedNodeWeigth: string;
@@ -941,7 +922,7 @@ export class SankeyDiagram implements IVisual {
     }
 
     private parseSettings(dataView: DataView): SankeyDiagramSettings {
-        let settings: SankeyDiagramSettings = SankeyDiagramSettings.parse<SankeyDiagramSettings>(dataView);
+        const settings: SankeyDiagramSettings = SankeyDiagramSettings.parse<SankeyDiagramSettings>(dataView);
 
         // settings.valueSourcesQuery = dataView.matrix.valueSources && dataView.matrix.valueSources[0].queryName;
 
@@ -976,11 +957,10 @@ export class SankeyDiagram implements IVisual {
 
     // tslint:disable-next-line: max-func-body-length
     private computePositions(sankeyDiagramDataView: SankeyDiagramDataView): void {
-        let maxXPosition: number,
-            maxColumn: SankeyDiagramColumn,
+        let maxColumn: SankeyDiagramColumn,
             columns: SankeyDiagramColumn[];
 
-        maxXPosition = SankeyDiagram.computeXPositions(sankeyDiagramDataView);
+        const maxXPosition: number = SankeyDiagram.computeXPositions(sankeyDiagramDataView);
 
         SankeyDiagram.sortNodesByX(sankeyDiagramDataView.nodes);
 
@@ -990,7 +970,7 @@ export class SankeyDiagram implements IVisual {
         let scaleStepCount: number = 0;
 
         let minWeigthShift: number = 0;
-        let minWeigthLink = _.minBy(sankeyDiagramDataView.links, "weigth");
+        const minWeigthLink = lodashMinBy(sankeyDiagramDataView.links, "weigth");
         if (minWeigthLink) {
             minWeigthShift = minWeigthLink.weigth;
         }
@@ -998,26 +978,26 @@ export class SankeyDiagram implements IVisual {
             minWeigthShift = 0;
         }
 
-        let minWeightInData: number = minWeigthShift;
+        const minWeightInData: number = minWeigthShift;
         minWeigthShift = Math.abs(minWeigthShift) + minWeight;
         let maxWeightInData: number = 0;
-        let maxWeigthLink = _.maxBy(sankeyDiagramDataView.links, "weigth");
+        const maxWeigthLink = lodashMaxBy(sankeyDiagramDataView.links, "weigth");
         if (maxWeigthLink) {
             maxWeightInData = maxWeigthLink.weigth;
         }
 
-        let minRangeOfScale: number = sankeyDiagramDataView.settings.scaleSettings.provideMinHeight ? SankeyDiagram.DefaultMinRangeOfScale : SankeyDiagram.MinRangeOfScale;
+        const minRangeOfScale: number = sankeyDiagramDataView.settings.scaleSettings.provideMinHeight ? SankeyDiagram.DefaultMinRangeOfScale : SankeyDiagram.MinRangeOfScale;
 
         while (minHeight <= SankeyDiagram.MinHeightOfNode && scaleStepCount < SankeyDiagram.ScaleStepLimit) {
             let weightScale: any;
 
             if (sankeyDiagramDataView.settings.scaleSettings.lnScale) {
-                weightScale = d3.scaleLog()
+                weightScale = d3ScaleLog()
                     .base(Math.E)
                     .domain([Math.exp(SankeyDiagram.MinDomainOfScale + scaleShift), Math.exp(SankeyDiagram.MaxDomainOfScale + scaleShift)])
                     .range([minRangeOfScale, SankeyDiagram.DefaultMaxRangeOfScale]);
             } else {
-                weightScale = d3.scaleLinear()
+                weightScale = d3ScaleLinear()
                     .domain([minWeightInData + scaleShift, maxWeightInData + scaleShift])
                     .range([minRangeOfScale, SankeyDiagram.DefaultMaxRangeOfScale]);
             }
@@ -1052,7 +1032,7 @@ export class SankeyDiagram implements IVisual {
             columns = this.getColumns(sankeyDiagramDataView.nodes);
             maxColumn = SankeyDiagram.getMaxColumn(columns);
 
-            minWeight = d3.min(sankeyDiagramDataView.nodes.filter((n) => Math.max(n.inputWeight, n.outputWeight) > 0).map((n) => Math.max(n.inputWeight, n.outputWeight)));
+            minWeight = d3Min(sankeyDiagramDataView.nodes.filter((n) => Math.max(n.inputWeight, n.outputWeight) > 0).map((n) => Math.max(n.inputWeight, n.outputWeight)));
             minWeight = minWeight || SankeyDiagram.DefaultWeightOfLink;
             sankeyDiagramDataView.settings._scale.y = this.getScaleByAxisY(maxColumn.sumValueOfNodes);
 
@@ -1086,7 +1066,7 @@ export class SankeyDiagram implements IVisual {
 
     private applySavedPositions(sankeyDiagramDataView: SankeyDiagramDataView) {
         // if size were changed shift positions of nodes
-        let viewPort: ViewportSize = sankeyDiagramDataView.settings._viewportSize;
+        const viewPort: ViewportSize = sankeyDiagramDataView.settings._viewportSize;
         let scaleHeight: number = 1;
         if (+viewPort.height !== this.viewport.height && viewPort.height && +viewPort.height !== 0) {
             scaleHeight = this.viewport.height / +viewPort.height;
@@ -1106,7 +1086,7 @@ export class SankeyDiagram implements IVisual {
 
     private computeBordersOfTheNode(sankeyDiagramDataView: SankeyDiagramDataView): void {
         sankeyDiagramDataView.nodes.forEach((node: SankeyDiagramNode) => {
-            let textHeight: number = textMeasurementService.estimateSvgTextHeight({
+            const textHeight: number = textMeasurementService.estimateSvgTextHeight({
                 text: node.label.formattedName,
                 fontFamily: this.textProperties.fontFamily,
                 fontSize: this.textProperties.fontSize
@@ -1153,14 +1133,14 @@ export class SankeyDiagram implements IVisual {
     }
 
     private static getUniqueLinks(links: SankeyDiagramLink[]) {
-        let unique = {};
+        const unique = {};
 
         links.forEach((link: SankeyDiagramLink) => {
             unique[link.source.label.name + link.destination.label.name + link.direction] = link;
         });
 
-        let newarray = [];
-        for (let key of Object.keys(unique)) {
+        const newarray = [];
+        for (const key of Object.keys(unique)) {
             newarray.push(unique[key]);
         }
 
@@ -1231,8 +1211,7 @@ export class SankeyDiagram implements IVisual {
     }
 
     public getColumns(nodes: SankeyDiagramNode[]): SankeyDiagramColumn[] {
-        let columns: SankeyDiagramColumn[] = [],
-            currentX: number = -Number.MAX_VALUE;
+        const columns: SankeyDiagramColumn[] = [];
 
         nodes.forEach((node: SankeyDiagramNode) => {
             if (!columns[node.x]) {
@@ -1303,7 +1282,7 @@ export class SankeyDiagram implements IVisual {
 
 
         columns.forEach(col => {
-            let sortedColumn = nodes
+            const sortedColumn = nodes
                 .slice(current, current + col.countOfNodes)
                 .sort((a, b) => {
                     let x, y;
@@ -1354,7 +1333,7 @@ export class SankeyDiagram implements IVisual {
 
             node.x *= scale.x;
 
-            let selfLinkHeight: number = d3.max(node.links.filter(l => l.direction === SankeyLinkDirrections.SelfLink).map(l => l.weigth));
+            let selfLinkHeight: number = d3Max(node.links.filter(l => l.direction === SankeyLinkDirrections.SelfLink).map(l => l.weigth));
 
             if (!selfLinkHeight) {
                 selfLinkHeight = 0;
@@ -1366,7 +1345,7 @@ export class SankeyDiagram implements IVisual {
             node.height = (Math.max(node.inputWeight, node.outputWeight, node.inputWeight + selfLinkHeight, node.outputWeight + selfLinkHeight)
             ) * scale.y;
 
-            let backwardPsudoNodeSpace = SankeyDiagram.BackwardPsudoNodeMargin + d3.max([node.backwardWeight, node.selftLinkWeight / 2]) * scale.y;
+            const backwardPsudoNodeSpace = SankeyDiagram.BackwardPsudoNodeMargin + d3Max([node.backwardWeight, node.selftLinkWeight / 2]) * scale.y;
 
             node.y = shiftByAxisY + offsetByY * index + backwardPsudoNodeSpace;
             shiftByAxisY += node.height;
@@ -1385,14 +1364,12 @@ export class SankeyDiagram implements IVisual {
         nodes.forEach((node: SankeyDiagramNode) => {
             node.links = SankeyDiagram.getUniqueLinks(node.links);
             node.links = node.links.sort((firstLink: SankeyDiagramLink, secondLink: SankeyDiagramLink) => {
-                let firstY: number,
-                    secondY: number;
 
-                firstY = firstLink.source === node
+                const firstY: number = firstLink.source === node
                     ? firstLink.destination.y
                     : firstLink.source.y;
 
-                secondY = secondLink.source === node
+                const secondY: number = secondLink.source === node
                     ? secondLink.destination.y
                     : secondLink.source.y;
 
@@ -1471,15 +1448,12 @@ export class SankeyDiagram implements IVisual {
     }
 
     private render(sankeyDiagramDataView: SankeyDiagramDataView): void {
-        let nodesSelection: Selection<SankeyDiagramNode>,
-            linksSelection: Selection<SankeyDiagramLink>;
-
-        linksSelection = this.renderLinks(sankeyDiagramDataView);
+        const linksSelection: Selection<SankeyDiagramLink> = this.renderLinks(sankeyDiagramDataView);
         this.renderLinkLabels(sankeyDiagramDataView);
 
         this.renderTooltip(linksSelection);
 
-        nodesSelection = this.renderNodes(sankeyDiagramDataView);
+        const nodesSelection: Selection<SankeyDiagramNode> = this.renderNodes(sankeyDiagramDataView);
 
         this.renderTooltip(nodesSelection);
 
@@ -1490,11 +1464,11 @@ export class SankeyDiagram implements IVisual {
 
     // tslint:disable-next-line: max-func-body-length
     private renderNodes(sankeyDiagramDataView: SankeyDiagramDataView): Selection<SankeyDiagramNode> {
-        let nodeElements: Selection<SankeyDiagramNode> = this.main
+        const nodeElements: Selection<SankeyDiagramNode> = this.main
             .select(SankeyDiagram.NodesSelector.selectorName)
             .selectAll(SankeyDiagram.NodeSelector.selectorName);
 
-        let nodesSelectionData = nodeElements
+        const nodesSelectionData = nodeElements
             .data(
                 sankeyDiagramDataView.nodes
                     .filter((node: SankeyDiagramNode) => {
@@ -1506,7 +1480,7 @@ export class SankeyDiagram implements IVisual {
             .exit()
             .remove();
 
-        let nodesEnterSelection: Selection<SankeyDiagramNode> = nodesSelectionData
+        const nodesEnterSelection: Selection<SankeyDiagramNode> = nodesSelectionData
             .enter()
             .append("g");
 
@@ -1518,7 +1492,7 @@ export class SankeyDiagram implements IVisual {
             .append("text")
             .classed(SankeyDiagram.NodeLabelSelector.className, true);
 
-        let nodesSelectionMerged = nodesEnterSelection.merge(nodeElements);
+        const nodesSelectionMerged = nodesEnterSelection.merge(nodeElements);
 
         nodesSelectionMerged
             .attr("transform", (node: SankeyDiagramNode) => {
@@ -1531,7 +1505,7 @@ export class SankeyDiagram implements IVisual {
             .style("fill", (node: SankeyDiagramNode) => node.fillColor)
             .style(
                 "stroke", (node: SankeyDiagramNode) => this.colorHelper.isHighContrast ? node.strokeColor :
-                    d3.rgb(node.fillColor)
+                    d3Rgb(node.fillColor)
                         .darker(SankeyDiagram.StrokeColorFactor)
                         .toString()
             )
@@ -1549,10 +1523,9 @@ export class SankeyDiagram implements IVisual {
             .style("font-family", this.textProperties.fontFamily)
             .style("font-size", this.textProperties.fontSize)
             .style("display", (node: SankeyDiagramNode) => {
-                let isNotVisibleLabel: boolean,
-                    labelPositionByAxisX: number = this.getCurrentPositionOfLabelByAxisX(node);
+                const labelPositionByAxisX: number = this.getCurrentPositionOfLabelByAxisX(node);
 
-                isNotVisibleLabel =
+                const isNotVisibleLabel: boolean =
                     (labelPositionByAxisX >= this.viewport.width ||
                         labelPositionByAxisX <= SankeyDiagram.MinSize ||
                         (node.height + SankeyDiagram.NodeMargin) < node.label.height) && !sankeyDiagramDataView.settings.labels.forceDisplay;
@@ -1584,98 +1557,80 @@ export class SankeyDiagram implements IVisual {
             });
 
         function dragstarted(node: SankeyDiagramNode) {
-            (<any>getEvent()).sourceEvent.stopPropagation();
+            event.stopPropagation();
         }
 
-        let minHeight: number = d3.min(sankeyDiagramDataView.links.map(l => l.height));
+        const minHeight: number = d3Min(sankeyDiagramDataView.links.map(l => l.height));
 
-        let sankeyVisual = this;
-        function dragged(node: SankeyDiagramNode) {
-            node.x = (getEvent()).x;
-            node.y = (getEvent()).y;
+        // eslint-disable-next-line
+        let self = this;
 
+        function dragged(event: DragEvent, node: SankeyDiagramNode) {
+            node.x = event.x;
+            node.y = event.y;
             if (node.x < 0) {
                 node.x = 0;
             }
-
             if (node.y < 0) {
                 node.y = 0;
             }
-
-            if (node.x + node.width > sankeyVisual.viewport.width) {
-                node.x = sankeyVisual.viewport.width - node.width;
+            if (node.x + node.width > self.viewport.width) {
+                node.x = self.viewport.width - node.width;
             }
-
-            if (node.y + node.height > sankeyVisual.viewport.height) {
-                node.y = sankeyVisual.viewport.height - node.height;
+            if (node.y + node.height > self.viewport.height) {
+                node.y = self.viewport.height - node.height;
             }
-
-            node.settings = {
-                x: node.x.toFixed(2),
-                y: node.y.toFixed(2),
-                name: node.label.name
+            node.settings = { x: node.x.toFixed(2), y: node.y.toFixed(2),name: node.label.name
             };
-
             // Update each link related with this node
             node.links.forEach((link: SankeyDiagramLink) => {
                 // select link svg element by ID generated in link creation as Source-Destination
-                d3.select(`#${SankeyDiagram.createLink(link, true)}`).attr(
+                d3Select(`#${SankeyDiagram.createLink(link, true)}`).attr(
                     // get updated path params based on actual positions of node
                     "d", (link: SankeyDiagramLink) => {
                         if (link.direction === SankeyLinkDirrections.Forward) {
-                            return sankeyVisual.getSvgPathForForwardLink(link);
+                            return self.getSvgPathForForwardLink(link);
                         }
                         if (link.direction === SankeyLinkDirrections.Backward) {
                             if (link.source.x + link.source.width > link.destination.x) {
-                                return sankeyVisual.getSvgPathForForwardLink(link);
+                                return self.getSvgPathForForwardLink(link);
                             }
-                            return sankeyVisual.getSvgPathForBackwardLink(link);
+                            return self.getSvgPathForBackwardLink(link);
                         }
                         if (link.direction === SankeyLinkDirrections.SelfLink) {
-                            return sankeyVisual.getSvgPathForSelfLink(link, minHeight);
+                            return self.getSvgPathForSelfLink(link, minHeight);
                         }
                     }
                 );
-                d3.select(`#${SankeyDiagram.createLink(link)}`).attr(
+                d3Select(`#${SankeyDiagram.createLink(link)}`).attr(
                     // get updated path params based on actual positions of node
                     "d", (link: SankeyDiagramLink) => {
                         if (link.direction === SankeyLinkDirrections.Forward) {
-                            return sankeyVisual.getSvgPathForForwardLink(link);
+                            return self.getSvgPathForForwardLink(link);
                         }
                         if (link.direction === SankeyLinkDirrections.Backward) {
                             if (link.source.x + link.source.width > link.destination.x) {
-                                return sankeyVisual.getSvgPathForForwardLink(link);
+                                return self.getSvgPathForForwardLink(link);
                             }
-                            return sankeyVisual.getSvgPathForBackwardLink(link);
+                            return self.getSvgPathForBackwardLink(link);
                         }
                         if (link.direction === SankeyLinkDirrections.SelfLink) {
-                            return sankeyVisual.getSvgPathForSelfLink(link, minHeight);
+                            return self.getSvgPathForSelfLink(link, minHeight);
                         }
                     }
                 );
             });
-
             // Translate the object on the actual moved point
-            d3.select(this).attr(
-                "transform", translate(node.x, node.y)
-            );
+            d3Select(this).attr("transform", translate(node.x, node.y));
         }
 
         function dragend(node: SankeyDiagramNode) {
-            sankeyVisual.saveNodePositions(sankeyVisual.dataView.nodes);
-            sankeyVisual.saveViewportSize();
+            self.saveNodePositions(self.dataView.nodes);
+            self.saveViewportSize();
         }
 
-        let drag = d3.drag()
-            .subject((node: SankeyDiagramNode) => {
-                return { x: node.x, y: node.y };
-            })
-            .on("start", dragstarted)
-            .on("drag", dragged)
-            .on("end", dragend);
-
+        const drag = d3Drag().on("start", dragstarted).on("drag", dragged).on("end", dragend);
         nodesSelectionMerged.call(drag);
-
         return nodesSelectionMerged;
     }
 
@@ -1699,12 +1654,12 @@ export class SankeyDiagram implements IVisual {
     }
 
     private saveNodePositions(nodes: SankeyDiagramNode[]): void {
-        let nodePositions: SankeyDiagramNodePositionSetting[] = [];
+        const nodePositions: SankeyDiagramNodePositionSetting[] = [];
         nodes.forEach((node: SankeyDiagramNode) => {
             if (node.height === 0) {
                 return;
             }
-            let settings: SankeyDiagramNodePositionSetting = <SankeyDiagramNodePositionSetting>{
+            const settings: SankeyDiagramNodePositionSetting = <SankeyDiagramNodePositionSetting>{
                 name: node.label.name,
                 x: node.x.toFixed(0),
                 y: node.y.toFixed(0)
@@ -1752,11 +1707,11 @@ export class SankeyDiagram implements IVisual {
     }
 
     private renderLinks(sankeyDiagramDataView: SankeyDiagramDataView): Selection<SankeyDiagramLink> {
-        let linksElements: Selection<SankeyDiagramLink> = this.main
+        const linksElements: Selection<SankeyDiagramLink> = this.main
             .select(SankeyDiagram.LinksSelector.selectorName)
             .selectAll(SankeyDiagram.LinkSelector.selectorName);
 
-        let linksSelectionData: UpdateSelection<SankeyDiagramLink> =
+        const linksSelectionData: UpdateSelection<SankeyDiagramLink> =
             linksElements
                 .data(
                     sankeyDiagramDataView.links.filter((link: SankeyDiagramLink) => {
@@ -1772,11 +1727,11 @@ export class SankeyDiagram implements IVisual {
             .exit()
             .remove();
 
-        let linksElementsEnter = linksSelectionData
+        const linksElementsEnter = linksSelectionData
             .enter()
             .append("path");
 
-        let linksElementsMerged = linksElementsEnter.merge(linksElements);
+        const linksElementsMerged = linksElementsEnter.merge(linksElements);
 
         linksElementsMerged
             .classed(SankeyDiagram.LinkSelector.className, true)
@@ -1784,7 +1739,7 @@ export class SankeyDiagram implements IVisual {
             .classed(SankeyDiagram.SelftLinkSelector.className, (link: SankeyDiagramLink) => link.direction === SankeyLinkDirrections.SelfLink);
 
 
-        let minHeight = d3.min(sankeyDiagramDataView.links.map(l => l.height));
+        const minHeight = d3Min(sankeyDiagramDataView.links.map(l => l.height));
 
         linksElementsMerged
             .attr(
@@ -1821,20 +1776,20 @@ export class SankeyDiagram implements IVisual {
     // tslint:disable-next-line: max-func-body-length
     private renderLinkLabels(sankeyDiagramDataView: SankeyDiagramDataView): void {
         // create labels on link as A - B : Value
-        let linkTextData: SankeyDiagramLink[] = sankeyDiagramDataView.links.filter((link: SankeyDiagramLink) => {
+        const linkTextData: SankeyDiagramLink[] = sankeyDiagramDataView.links.filter((link: SankeyDiagramLink) => {
             return link.height > SankeyDiagram.MinSize && this.dataView.settings.linkLabels.show;
         });
 
-        let linkArrowData: SankeyDiagramLink[] = sankeyDiagramDataView.links.filter((link: SankeyDiagramLink) => {
+        const linkArrowData: SankeyDiagramLink[] = sankeyDiagramDataView.links.filter((link: SankeyDiagramLink) => {
             return link.height > SankeyDiagram.MinSize && link.direction !== SankeyLinkDirrections.SelfLink;
         });
 
         // add defs element to svg
-        let svgDefs: Selection<any> = this.root
+        const svgDefs: Selection<any> = this.root
             .selectAll("defs");
 
-        let svgDefsSelectionData: UpdateSelection<Number> = svgDefs.data([1]);
-        let svgDefsSelectionEnter = svgDefsSelectionData
+        const svgDefsSelectionData: UpdateSelection<number> = svgDefs.data([1]);
+        const svgDefsSelectionEnter = svgDefsSelectionData
             .enter()
             .append("defs");
 
@@ -1842,24 +1797,24 @@ export class SankeyDiagram implements IVisual {
             .exit()
             .remove();
 
-        let svgDefsSelectionMerged = svgDefsSelectionEnter.merge(svgDefs);
+        const svgDefsSelectionMerged = svgDefsSelectionEnter.merge(svgDefs);
 
-        let singleDefsElement: Selection<any> = d3.select(svgDefsSelectionMerged.node());
+        const singleDefsElement: Selection<any> = d3Select(svgDefsSelectionMerged.node());
 
         // add text path for lables
-        let linkLabelPaths: Selection<any> = singleDefsElement.selectAll(SankeyDiagram.LinkLabelPathsSelector.selectorName);
+        const linkLabelPaths: Selection<any> = singleDefsElement.selectAll(SankeyDiagram.LinkLabelPathsSelector.selectorName);
 
-        let linkLabelPathsSelectionData: UpdateSelection<SankeyDiagramLink> = linkLabelPaths.data(linkArrowData);
+        const linkLabelPathsSelectionData: UpdateSelection<SankeyDiagramLink> = linkLabelPaths.data(linkArrowData);
 
         linkLabelPathsSelectionData
             .exit()
             .remove();
 
-        let linkLabelPathsSelectionEnter = linkLabelPathsSelectionData
+        const linkLabelPathsSelectionEnter = linkLabelPathsSelectionData
             .enter()
             .append("path");
 
-        let linkLabelPathsSelectionMerged = linkLabelPathsSelectionEnter.merge(linkLabelPaths);
+        const linkLabelPathsSelectionMerged = linkLabelPathsSelectionEnter.merge(linkLabelPaths);
 
         linkLabelPathsSelectionMerged
             .classed(SankeyDiagram.LinkLabelPathsSelector.className, true)
@@ -1877,41 +1832,41 @@ export class SankeyDiagram implements IVisual {
             );
 
         // add text by using paths from defs
-        let linkLabelTexts: Selection<any> = this.main
+        const linkLabelTexts: Selection<any> = this.main
             .select(SankeyDiagram.LinksSelector.selectorName)
             .selectAll(SankeyDiagram.LinkLabelTextsSelector.selectorName);
 
-        let linkLabelTextSelectionData: UpdateSelection<SankeyDiagramLink> = linkLabelTexts
+        const linkLabelTextSelectionData: UpdateSelection<SankeyDiagramLink> = linkLabelTexts
             .data(linkTextData);
 
         linkLabelTextSelectionData
             .exit()
             .remove();
 
-        let linkLabelTextSelectionEnter = linkLabelTextSelectionData
+        const linkLabelTextSelectionEnter = linkLabelTextSelectionData
             .enter()
             .append("text");
 
-        let linkLabelTextSelectionMerged = linkLabelTextSelectionEnter.merge(linkLabelTexts);
+        const linkLabelTextSelectionMerged = linkLabelTextSelectionEnter.merge(linkLabelTexts);
 
         linkLabelTextSelectionMerged
             .attr("text-anchor", "middle")
             .classed(SankeyDiagram.LinkLabelTextsSelector.className, true);
 
-        let textPathSelection: UpdateSelection<SankeyDiagramLink> = linkLabelTextSelectionMerged
+        const textPathSelection: UpdateSelection<SankeyDiagramLink> = linkLabelTextSelectionMerged
             .selectAll("textPath");
 
-        let textPathSelectionData = textPathSelection.data(data => [data]);
+        const textPathSelectionData = textPathSelection.data(data => [data]);
 
         textPathSelectionData
             .exit()
             .remove();
 
-        let textPathSelectionEnter = textPathSelectionData
+            const textPathSelectionEnter = textPathSelectionData
             .enter()
             .append("textPath");
 
-        let textPathSelectionMerged = textPathSelectionEnter.merge(textPathSelection);
+        const textPathSelectionMerged = textPathSelectionEnter.merge(textPathSelection);
 
         textPathSelectionMerged
             .attr("startOffset", "50%")
@@ -1927,13 +1882,7 @@ export class SankeyDiagram implements IVisual {
     }
 
     private getLinkLabelSvgPath(link: SankeyDiagramLink): string {
-        let x0: number,
-            x1: number,
-            xi: (t: number) => number,
-            x2: number,
-            x3: number,
-            y0: number,
-            y1: number;
+        let x0: number, x1: number;
 
         if (link.destination.x < link.source.x) {
             x0 = link.source.x - 10;
@@ -1943,12 +1892,12 @@ export class SankeyDiagram implements IVisual {
             x1 = link.destination.x - 10;
         }
 
-        xi = d3.interpolateNumber(x0, x1);
-        x2 = xi(this.curvatureOfLinks);
-        x3 = xi(1 - this.curvatureOfLinks);
+        const xi: (t: number) => number = d3InterpolateNumber(x0, x1);
+        const x2: number = xi(this.curvatureOfLinks);
+        const x3: number = xi(1 - this.curvatureOfLinks);
 
-        y0 = link.source.y - (link.direction === SankeyLinkDirrections.Backward ? link.height + SankeyDiagram.NodeAndBackwardLinkDistance : 0) + link.dySource + link.height / SankeyDiagram.MiddleFactor;
-        y1 = link.destination.y - (link.direction === SankeyLinkDirrections.Backward ? link.height + SankeyDiagram.NodeAndBackwardLinkDistance : 0) + (link.dyDestination || 0) + link.height / SankeyDiagram.MiddleFactor;
+        const y0: number = link.source.y - (link.direction === SankeyLinkDirrections.Backward ? link.height + SankeyDiagram.NodeAndBackwardLinkDistance : 0) + link.dySource + link.height / SankeyDiagram.MiddleFactor;
+        const y1: number = link.destination.y - (link.direction === SankeyLinkDirrections.Backward ? link.height + SankeyDiagram.NodeAndBackwardLinkDistance : 0) + (link.dyDestination || 0) + link.height / SankeyDiagram.MiddleFactor;
 
         return `M ${x0} ${y0} C ${x2} ${y0}, ${x3} ${y1}, ${x1} ${y1}`;
     }
@@ -1965,13 +1914,10 @@ export class SankeyDiagram implements IVisual {
             fixedLinkHeight = Math.min(link.destination.width, minHeight);
         }
 
-        let linkKneeSize: number = Math.min(link.destination.width, minHeight);
+        const linkKneeSize: number = Math.min(link.destination.width, minHeight);
 
         let x0: number,
             x1: number,
-            xi: (t: number) => number,
-            x2: number,
-            x3: number,
             y0: number,
             y1: number;
 
@@ -2083,8 +2029,8 @@ export class SankeyDiagram implements IVisual {
         const distanceBetweenLinks: number = 3;
         const distanceFromNodeToLinks: number = 5;
 
-        let fixedLinkHeight = link.height - distanceBetweenLinks;
-        let linkKneeSize: number = link.height;
+        const fixedLinkHeight = link.height - distanceBetweenLinks;
+        const linkKneeSize: number = link.height;
 
         let x0: number,
             x1: number,
@@ -2105,7 +2051,7 @@ export class SankeyDiagram implements IVisual {
         // drawing area as combination of 4 lines in one path element of svg to fill this area with required color
 
         // upper border of link
-        xi = d3.interpolateNumber(x0, x1);
+        xi = d3InterpolateNumber(x0, x1);
         x2 = xi(this.curvatureOfLinks);
         x3 = xi(1 - this.curvatureOfLinks);
         y0 = link.source.y - (link.height + SankeyDiagram.NodeAndBackwardLinkDistance) + link.dySource + link.height / SankeyDiagram.MiddleFactor - link.height / 2;
@@ -2143,7 +2089,7 @@ export class SankeyDiagram implements IVisual {
         pathParams += `L ${link.destination.x + link.destination.width + distanceFromNodeToLinks} ${link.destination.y - SankeyDiagram.NodeAndBackwardLinkDistance}`;
 
         // bottom border of link
-        xi = d3.interpolateNumber(x0, x1);
+        xi = d3InterpolateNumber(x0, x1);
         x2 = xi(this.curvatureOfLinks);
         x3 = xi(1 - this.curvatureOfLinks);
         y0 = link.source.y - (link.height + SankeyDiagram.NodeAndBackwardLinkDistance) + link.dySource + link.height / SankeyDiagram.MiddleFactor + link.height / 2;
@@ -2200,7 +2146,7 @@ export class SankeyDiagram implements IVisual {
 
         // drawing area as combination of 4 lines in one path element of svg to fill this area with required color
         // upper border of link
-        xi = d3.interpolateNumber(x0, x1);
+        xi = d3InterpolateNumber(x0, x1);
         x2 = xi(this.curvatureOfLinks);
         x3 = xi(1 - this.curvatureOfLinks);
         y0 = link.source.y + link.dySource + link.height / SankeyDiagram.MiddleFactor - (link.height - distanceBetweenLinks) / 2;
@@ -2215,7 +2161,7 @@ export class SankeyDiagram implements IVisual {
         pathParams += ` L ${x1} ${y0}`;
 
         // bottom border of link
-        xi = d3.interpolateNumber(x0, x1);
+        xi = d3InterpolateNumber(x0, x1);
         x2 = xi(this.curvatureOfLinks);
         x3 = xi(1 - this.curvatureOfLinks);
         y0 = link.source.y + link.dySource + (link.height - distanceBetweenLinks) / SankeyDiagram.MiddleFactor + (link.height - distanceBetweenLinks) / 2;
@@ -2234,11 +2180,15 @@ export class SankeyDiagram implements IVisual {
     }
 
     private renderTooltip(selection: Selection<SankeyDiagramNode | SankeyDiagramLink>): void {
+        if (!this.tooltipServiceWrapper) {
+            return;
+        }
+
         this.tooltipServiceWrapper.addTooltip(
             selection,
-            (tooltipEvent: TooltipEventArgs<TooltipEnabledDataPoint>) => {
-                return tooltipEvent.data.tooltipInfo;
-            });
+            (data: TooltipEnabledDataPoint) => data.tooltipInfo,
+            (data: SankeyDiagramNode | SankeyDiagramLink) => data.identity
+        );
     }
 
     private updateSelectionState(